Update reporting requirements to be related interventions
