+++ EquiTrack/reports/serializers/v2.py
@@ -1,5 +1,3 @@
-from operator import itemgetter
-
 from django.db import transaction
 from django.utils.translation import ugettext as _
 
@@ -288,155 +286,3 @@ class ReportingRequirementSerializer(serializers.ModelSerializer):
     class Meta:
         model = ReportingRequirement
         fields = ("id", "start_date", "end_date", "due_date", "description", )
-
-
-class IndicatorReportingRequirementSerializer(serializers.ModelSerializer):
-    report_type = serializers.ChoiceField(
-        choices=ReportingRequirement.TYPE_CHOICES
-    )
-    reporting_requirements = ReportingRequirementSerializer(many=True)
-
-    class Meta:
-        model = AppliedIndicator
-        fields = ("id", "reporting_requirements", "report_type", )
-        read_only_fields = ("id", )
-
-    def _validate_qpr(self, intervention, requirements):
-        # Ensure that the first reporting requirement start date
-        # is on or after PD start date
-        if requirements[0]["start_date"] < intervention.start:
-            raise serializers.ValidationError({
-                "reporting_requirements": {
-                    "start_date": _(
-                        "Start date needs to be on or after PD start date."
-                    )
-                }
-            })
-
-        # Ensure start date is after previous end date
-        for i in range(1, len(requirements)):
-            if requirements[i]["start_date"] <= requirements[i-1]["end_date"]:
-                raise serializers.ValidationError({
-                    "reporting_requirements": {
-                        "start_date": _(
-                            "Start date needs to be after previous end date."
-                        )
-                    }
-                })
-
-    def _validate_hr(self, indicator):
-        # Ensure intervention is HF or cluster
-        if not indicator.is_high_frequency and not indicator.cluster_indicator_id:
-            raise serializers.ValidationError(
-                _("Indicator needs to be either cluster or high frequency.")
-            )
-
-    def _merge_data(self, data):
-        current_reqs = ReportingRequirement.objects.values(
-            "id",
-            "start_date",
-            "end_date",
-            "due_date",
-        ).filter(
-            applied_indicator=self.indicator,
-            report_type=data["report_type"],
-        )
-
-        current_reqs_dict = {}
-        for r in current_reqs:
-            current_reqs_dict[r["id"]] = r
-
-        report_type = data["report_type"]
-        for r in data["reporting_requirements"]:
-            if r.get("id") in current_reqs_dict:
-                current_reqs_dict.pop(r["id"])
-            r["applied_indicator"] = self.indicator
-            r["report_type"] = report_type
-            if report_type == ReportingRequirement.TYPE_HR:
-                r["end_date"] = r["due_date"]
-                r["start_date"] = None
-                r["description"] = ""
-            elif report_type == ReportingRequirement.TYPE_SPECIAL:
-                r["start_date"] = None
-                r["end_date"] = r["due_date"]
-
-        # We need all reporting requirements in end date order
-        merged_reqs = list(current_reqs_dict.values()) + data["reporting_requirements"]
-        data["reporting_requirements"] = sorted(
-            merged_reqs,
-            key=itemgetter("end_date")
-        )
-        return data
-
-    def run_validation(self, initial_data):
-        serializer = self.fields["reporting_requirements"].child
-        report_type = initial_data.get("report_type")
-        if report_type == ReportingRequirement.TYPE_QPR:
-            serializer.fields["start_date"].required = True
-            serializer.fields["end_date"].required = True
-        elif report_type == ReportingRequirement.TYPE_SPECIAL:
-            serializer.fields["description"].required = True
-        return super().run_validation(initial_data)
-
-    def validate(self, data):
-        """The first reporting requirement's start date needs to be
-        on or after the PD start date.
-        Subsequent reporting requirements start date needs to be after the
-        previous reporting requirement end date.
-        """
-        pk = self.initial_data.get("id")
-        try:
-            self.indicator = self.Meta.model.objects.get(pk=pk)
-        except self.Meta.model.DoesNotExist:
-            raise serializers.ValidationError({
-                "id": _("Invalid indicator id.")
-            })
-
-        # Only able to change reporting requirements when PD
-        # is in amendment status
-        intervention = self.indicator.lower_result.result_link.intervention
-        if intervention.status not in [intervention.DRAFT]:
-            raise serializers.ValidationError(
-                _("Changes not allowed when PD not in amendment state.")
-            )
-        if not intervention.start:
-            raise serializers.ValidationError(
-                _("PD needs to have a start date.")
-            )
-
-        # Validate reporting requirements first
-        if not len(data["reporting_requirements"]):
-            raise serializers.ValidationError({
-                "reporting_requirements": _("This field cannot be empty.")
-            })
-
-        self._merge_data(data)
-
-        if data["report_type"] == ReportingRequirement.TYPE_QPR:
-            self._validate_qpr(intervention, data["reporting_requirements"])
-        elif data["report_type"] == ReportingRequirement.TYPE_HR:
-            self._validate_hr(self.indicator)
-
-        return data
-
-    def create(self, validated_data):
-        current_reqs = ReportingRequirement.objects.values_list(
-            "id",
-            flat=True
-        ).filter(
-            applied_indicator=self.indicator,
-            report_type=validated_data["report_type"]
-        )
-        new_reqs = [
-            r["id"] for r in validated_data["reporting_requirements"]
-            if "id" in r
-        ]
-        delete_reqs = [r for r in current_reqs if r not in new_reqs]
-        ReportingRequirement.objects.filter(id__in=delete_reqs).delete()
-        for r in validated_data["reporting_requirements"]:
-            if r.get("id"):
-                pk = r.pop("id")
-                ReportingRequirement.objects.filter(pk=pk).update(**r)
-            else:
-                ReportingRequirement.objects.create(**r)
-        return self.indicator
