+++ EquiTrack/reports/tests/test_serializers.py
@@ -1,12 +1,9 @@
 from __future__ import unicode_literals
 
-import datetime
-
 from rest_framework.exceptions import ValidationError
 
 from EquiTrack.tests.cases import BaseTenantTestCase
 from locations.tests.factories import LocationFactory
-from partners.models import Intervention
 from partners.tests.factories import (
     InterventionFactory,
     InterventionResultLinkFactory,
@@ -15,12 +12,10 @@ from reports.models import (
     AppliedIndicator,
     IndicatorBlueprint,
     LowerResult,
-    ReportingRequirement,
 )
 from reports.serializers.v2 import (
     AppliedIndicatorSerializer,
     DisaggregationSerializer,
-    IndicatorReportingRequirementSerializer,
     LowerResultCUSerializer,
     LowerResultSimpleCUSerializer,
 )
@@ -30,7 +25,6 @@ from reports.tests.factories import (
     DisaggregationValueFactory,
     IndicatorBlueprintFactory,
     LowerResultFactory,
-    ReportingRequirementFactory,
     SectorFactory,
 )
 
@@ -307,392 +301,3 @@ class TestLowerResultCUSerializer(BaseTenantTestCase):
         self.assertIsInstance(lower_result_updated, LowerResult)
         self.assertEqual(lower_result_updated.name, "LL Name")
         self.assertTrue(lower_result_qs.exists())
-
-
-class TestIndicatorReportingRequirementSerializer(BaseTenantTestCase):
-    @classmethod
-    def setUpTestData(cls):
-        cls.intervention = InterventionFactory(start=datetime.date(2001, 1, 1))
-        cls.result_link = InterventionResultLinkFactory(
-            intervention=cls.intervention
-        )
-        cls.lower_result = LowerResultFactory(result_link=cls.result_link)
-        cls.indicator = AppliedIndicatorFactory(lower_result=cls.lower_result)
-
-    def test_validation_invalid_indicator(self):
-        data = {
-            "id": 404,
-            "report_type": ReportingRequirement.TYPE_QPR,
-            "reporting_requirements": []
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(serializer.errors['id'], ['Invalid indicator id.'])
-
-    def test_validation_missing_indicator(self):
-        data = {
-            "report_type": ReportingRequirement.TYPE_QPR,
-            "reporting_requirements": []
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(serializer.errors['id'], ['Invalid indicator id.'])
-
-    def test_validation_invalid_report_type(self):
-        data = {
-            "id": self.indicator.pk,
-            "report_type": "wrong",
-            "reporting_requirements": []
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors['report_type'],
-            ['"wrong" is not a valid choice.']
-        )
-
-    def test_validation_missing_report_type(self):
-        data = {
-            "id": self.indicator.pk,
-            "reporting_requirements": []
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors['report_type'],
-            ['This field is required.']
-        )
-
-    def test_validation_pd_status(self):
-        intervention = InterventionFactory(status=Intervention.CLOSED)
-        result_link = InterventionResultLinkFactory(intervention=intervention)
-        lower_result = LowerResultFactory(result_link=result_link)
-        indicator = AppliedIndicatorFactory(lower_result=lower_result)
-        data = {
-            "id": indicator.pk,
-            "report_type": ReportingRequirement.TYPE_QPR,
-            "reporting_requirements": []
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors['non_field_errors'],
-            ['Changes not allowed when PD not in amendment state.']
-        )
-
-    def test_validation_pd_has_no_start(self):
-        intervention = InterventionFactory()
-        result_link = InterventionResultLinkFactory(intervention=intervention)
-        lower_result = LowerResultFactory(result_link=result_link)
-        indicator = AppliedIndicatorFactory(lower_result=lower_result)
-        data = {
-            "id": indicator.pk,
-            "report_type": ReportingRequirement.TYPE_QPR,
-            "reporting_requirements": []
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors['non_field_errors'],
-            ['PD needs to have a start date.']
-        )
-
-    def test_validation_empty_reporting_requirements(self):
-        data = {
-            "id": self.indicator.pk,
-            "report_type": ReportingRequirement.TYPE_QPR,
-            "reporting_requirements": []
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors['reporting_requirements'],
-            ['This field cannot be empty.']
-        )
-
-    def test_validation_missing_reporting_requirements(self):
-        data = {
-            "id": self.indicator.pk,
-            "report_type": ReportingRequirement.TYPE_QPR,
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors['reporting_requirements'],
-            ['This field is required.']
-        )
-
-    def test_validation_qpr_missing_fields(self):
-        data = {
-            "id": self.indicator.pk,
-            "report_type": ReportingRequirement.TYPE_QPR,
-            "reporting_requirements": [{
-                "end_date": datetime.date(2001, 3, 31),
-                "due_date": datetime.date(2001, 4, 15),
-            }]
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors['reporting_requirements'],
-            [{"start_date": ['This field is required.']}]
-        )
-
-    def test_validation_qpr_start_early(self):
-        data = {
-            "id": self.indicator.pk,
-            "report_type": ReportingRequirement.TYPE_QPR,
-            "reporting_requirements": [{
-                "start_date": datetime.date(2000, 1, 1),
-                "end_date": datetime.date(2001, 3, 31),
-                "due_date": datetime.date(2001, 4, 15),
-            }]
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors['reporting_requirements'],
-            {"start_date": 'Start date needs to be on or after PD start date.'}
-        )
-
-    def test_validation_qpr_end_before_start(self):
-        data = {
-            "id": self.indicator.pk,
-            "report_type": ReportingRequirement.TYPE_QPR,
-            "reporting_requirements": [{
-                "start_date": datetime.date(2001, 1, 1),
-                "end_date": datetime.date(2001, 3, 31),
-                "due_date": datetime.date(2001, 4, 15),
-            }, {
-                "start_date": datetime.date(2001, 2, 1),
-                "end_date": datetime.date(2001, 4, 30),
-                "due_date": datetime.date(2001, 5, 15),
-            }]
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors['reporting_requirements'],
-            {"start_date": 'Start date needs to be after previous end date.'}
-        )
-
-    def test_validation_qpr(self):
-        data = {
-            "id": self.indicator.pk,
-            "report_type": ReportingRequirement.TYPE_QPR,
-            "reporting_requirements": [{
-                "start_date": datetime.date(2001, 1, 1),
-                "end_date": datetime.date(2001, 3, 31),
-                "due_date": datetime.date(2001, 4, 15),
-            }, {
-                "start_date": datetime.date(2001, 4, 1),
-                "end_date": datetime.date(2001, 5, 31),
-                "due_date": datetime.date(2001, 5, 15),
-            }]
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertTrue(serializer.is_valid())
-
-    def test_validation_hr_missing_fields(self):
-        data = {
-            "id": self.indicator.pk,
-            "report_type": ReportingRequirement.TYPE_HR,
-            "reporting_requirements": [{
-                "start_date": datetime.date(2001, 4, 15),
-            }]
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors['reporting_requirements'],
-            [{"due_date": ['This field is required.']}]
-        )
-
-    def test_validation_hr_indicator_invalid(self):
-        self.assertFalse(self.indicator.is_high_frequency)
-        data = {
-            "id": self.indicator.pk,
-            "report_type": ReportingRequirement.TYPE_HR,
-            "reporting_requirements": [
-                {"due_date": datetime.date(2001, 4, 15)},
-                {"due_date": datetime.date(2001, 5, 15)}
-            ]
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors['non_field_errors'],
-            ["Indicator needs to be either cluster or high frequency."]
-        )
-
-    def test_validation_hr(self):
-        indicator = AppliedIndicatorFactory(
-            is_high_frequency=True,
-            lower_result=self.lower_result
-        )
-        data = {
-            "id": indicator.pk,
-            "report_type": ReportingRequirement.TYPE_HR,
-            "reporting_requirements": [
-                {"due_date": datetime.date(2001, 4, 15)},
-                {"due_date": datetime.date(2001, 5, 15)}
-            ]
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertTrue(serializer.is_valid())
-
-    def test_validation_special_missing_fields(self):
-        data = {
-            "id": self.indicator.pk,
-            "report_type": ReportingRequirement.TYPE_SPECIAL,
-            "reporting_requirements": [{
-                "due_date": datetime.date(2001, 4, 15),
-            }]
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors['reporting_requirements'],
-            [{"description": ['This field is required.']}]
-        )
-
-    def test_validation_special_description_long(self):
-        data = {
-            "id": self.indicator.pk,
-            "report_type": ReportingRequirement.TYPE_SPECIAL,
-            "reporting_requirements": [{
-                "due_date": datetime.date(2001, 4, 15),
-                "description": "long"*256
-            }]
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors['reporting_requirements'],
-            [{
-                "description": [
-                    "Ensure this field has no more than 256 characters."
-                ]
-            }]
-        )
-
-    def test_validation_special(self):
-        data = {
-            "id": self.indicator.pk,
-            "report_type": ReportingRequirement.TYPE_SPECIAL,
-            "reporting_requirements": [{
-                "due_date": datetime.date(2001, 4, 15),
-                "description": "some description goes here"
-            }]
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertTrue(serializer.is_valid())
-
-    def test_create_qpr(self):
-        """Creating new qpr reporting requirements
-
-        When none currently existing
-        """
-        requirement_qs = ReportingRequirement.objects.filter(
-            applied_indicator=self.indicator,
-            report_type=ReportingRequirement.TYPE_QPR,
-        )
-        init_count = requirement_qs.count()
-        data = {
-            "id": self.indicator.pk,
-            "report_type": ReportingRequirement.TYPE_QPR,
-            "reporting_requirements": [{
-                "start_date": datetime.date(2001, 1, 1),
-                "end_date": datetime.date(2001, 3, 31),
-                "due_date": datetime.date(2001, 4, 15),
-            }, {
-                "start_date": datetime.date(2001, 4, 1),
-                "end_date": datetime.date(2001, 5, 31),
-                "due_date": datetime.date(2001, 5, 15),
-            }]
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertTrue(serializer.is_valid())
-        serializer.create(serializer.validated_data)
-        self.assertEqual(requirement_qs.count(), init_count + 2)
-
-    def test_create_hr(self):
-        """Creating new hr reporting requirements
-
-        When none currently existing
-        """
-        indicator = AppliedIndicatorFactory(
-            is_high_frequency=True,
-            lower_result=self.lower_result
-        )
-        requirement_qs = ReportingRequirement.objects.filter(
-            applied_indicator=indicator,
-            report_type=ReportingRequirement.TYPE_HR,
-        )
-        init_count = requirement_qs.count()
-        data = {
-            "id": indicator.pk,
-            "report_type": ReportingRequirement.TYPE_HR,
-            "reporting_requirements": [
-                {"due_date": datetime.date(2001, 4, 15)},
-                {"due_date": datetime.date(2001, 5, 15)}
-            ]
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertTrue(serializer.is_valid())
-        serializer.create(serializer.validated_data)
-        self.assertEqual(requirement_qs.count(), init_count + 2)
-
-    def test_create_special(self):
-        """Creating new special reporting requirements
-
-        When none currently existing"""
-        requirement_qs = ReportingRequirement.objects.filter(
-            applied_indicator=self.indicator,
-            report_type=ReportingRequirement.TYPE_SPECIAL,
-        )
-        init_count = requirement_qs.count()
-        data = {
-            "id": self.indicator.pk,
-            "report_type": ReportingRequirement.TYPE_SPECIAL,
-            "reporting_requirements": [{
-                "due_date": datetime.date(2001, 4, 15),
-                "description": "some description goes here"
-            }]
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertTrue(serializer.is_valid())
-        serializer.create(serializer.validated_data)
-        self.assertEqual(requirement_qs.count(), init_count + 1)
-
-    def test_update_qpr(self):
-        """Updating existing qpr reporting requirements"""
-        report_type = ReportingRequirement.TYPE_QPR
-        requirement = ReportingRequirementFactory(
-            applied_indicator=self.indicator,
-            report_type=report_type,
-            start_date=datetime.date(2001, 1, 3),
-        )
-        requirement_qs = ReportingRequirement.objects.filter(
-            applied_indicator=self.indicator,
-            report_type=ReportingRequirement.TYPE_QPR,
-        )
-        init_count = requirement_qs.count()
-        data = {
-            "id": self.indicator.pk,
-            "report_type": ReportingRequirement.TYPE_QPR,
-            "reporting_requirements": [{
-                "id": requirement.pk,
-                "start_date": datetime.date(2001, 1, 1),
-                "end_date": datetime.date(2001, 3, 31),
-                "due_date": datetime.date(2001, 4, 15),
-            }]
-        }
-        serializer = IndicatorReportingRequirementSerializer(data=data)
-        self.assertTrue(serializer.is_valid())
-        serializer.create(serializer.validated_data)
-        self.assertEqual(requirement_qs.count(), init_count)
-        req_updated = ReportingRequirement.objects.get(pk=requirement.pk)
-        self.assertEqual(req_updated.start_date, datetime.date(2001, 1, 1))
-        self.assertEqual(req_updated.end_date, datetime.date(2001, 3, 31))
-        self.assertEqual(req_updated.due_date, datetime.date(2001, 4, 15))
